[security] Fix vulnerability in GH action
Under some very specific circumstances, if a release PR was made to the
repo with some special characters, it could allow someone to execute
shell code on the GH action environment.

Not much of a big deal, because it would require that the shell command
be part of the branch name, along with some extra markup to avoid
clashing with the action code, but still, better to be safe than sorry.

CORESDK-2545